refactor: optmize the project structure (#1028)
* feat!: Add a guild id property for the member object

* refactor: optimize project structure

* revert: stuff from other branch

* refactor: move missing from stub files

* fix: resolve new circular imports
