+++ interactions/__init__.py
@@ -12,3 +12,4 @@ To see the documentation, please head over to the link here:
 from .client import *  # noqa: F401 F403 isort: skip
 from .api import *  # noqa: F401 F403
 from .base import *  # noqa: F401 F403
+from .utils import *  # noqa: F401 F403
