+++ interactions/api/models/__init__.py
@@ -5,7 +5,6 @@ This section of the library maintains
 and stores all of the data defining
 models for dispatched Gateway events.
 """
-from .attrs_utils import *  # noqa: F401 F403
 from .audit_log import *  # noqa: F401 F403
 from .channel import *  # noqa: F401 F403
 from .emoji import *  # noqa: F401 F403
