+++ interactions/api/models/audit_log.py
@@ -1,7 +1,7 @@
 from enum import IntEnum
 from typing import TYPE_CHECKING, List, Optional, TypeVar
 
-from .attrs_utils import DictSerializerMixin, convert_list, define, field
+from ...utils.attrs_utils import DictSerializerMixin, convert_list, define, field
 from .channel import Channel
 from .misc import Snowflake
 from .user import User
