+++ interactions/api/models/member.py
@@ -1,9 +1,10 @@
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, List, Optional, Union
 
-from ...client.models.utils import search_iterable
+from ...utils.attrs_utils import ClientSerializerMixin, convert_int, convert_list, define, field
+from ...utils.missing import MISSING
+from ...utils.utils import search_iterable
 from ..error import LibraryException
-from .attrs_utils import MISSING, ClientSerializerMixin, convert_int, convert_list, define, field
 from .channel import Channel
 from .flags import Permissions
 from .misc import AllowedMentions, File, IDMixin, Snowflake
