+++ interactions/api/models/misc.py
@@ -15,8 +15,9 @@ from os.path import basename
 from typing import List, Optional, Union
 
 from ...base import get_logger
+from ...utils.attrs_utils import DictSerializerMixin, convert_list, define, field
+from ...utils.missing import MISSING
 from ..error import LibraryException
-from .attrs_utils import MISSING, DictSerializerMixin, convert_list, define, field
 
 __all__ = (
     "AutoModKeywordPresetTypes",
