+++ interactions/api/models/presence.py
@@ -2,7 +2,7 @@ import time
 from enum import IntEnum
 from typing import Any, List, Optional
 
-from .attrs_utils import DictSerializerMixin, convert_list, define, field
+from ...utils.attrs_utils import DictSerializerMixin, convert_list, define, field
 from .emoji import Emoji
 from .flags import StatusType
 from .misc import Snowflake
