+++ interactions/api/models/role.py
@@ -1,7 +1,8 @@
 from typing import TYPE_CHECKING, Any, List, Optional, Union
 
+from ...utils.attrs_utils import ClientSerializerMixin, DictSerializerMixin, define, field
+from ...utils.missing import MISSING
 from ..error import LibraryException
-from .attrs_utils import MISSING, ClientSerializerMixin, DictSerializerMixin, define, field
 from .misc import IDMixin, Image, Snowflake
 
 if TYPE_CHECKING:
