+++ interactions/client/__init__.py
@@ -8,5 +8,4 @@ from .bot import *  # noqa: F401 F403
 from .context import *  # noqa: F401 F403
 from .decor import *  # noqa: F401 F403
 from .enums import *  # noqa: F401 F403
-from .get import *  # noqa: F401 F403
 from .models import *  # noqa: F401 F403
