+++ interactions/client/models/__init__.py
@@ -8,4 +8,3 @@ models for interaction creation events.
 from .command import *  # noqa: F401 F403
 from .component import *  # noqa: F401 F403
 from .misc import *  # noqa: F401 F403
-from .utils import *  # noqa: F401 F403
