+++ interactions/utils/attrs_utils.py
@@ -1,36 +1,15 @@
 from copy import deepcopy
 from functools import wraps
-from typing import ClassVar, Dict, Mapping, Optional, Tuple
+from typing import Dict, Mapping, Optional, Tuple
 
 import attrs
 
-__all__ = ("MISSING", "DictSerializerMixin", "ClientSerializerMixin")
+from .missing import MISSING
 
-
-class _Missing:
-    """A sentinel object for places where None is a valid value"""
-
-    _instance: ClassVar["_Missing"] = None
-
-    def __new__(cls):
-        if not isinstance(cls._instance, cls):
-            cls._instance = object.__new__(cls)
-        return cls._instance
-
-    def __eq__(self, other):
-        return self.__class__ is other.__class__
-
-    def __repr__(self):
-        return "<interactions.MISSING>"
-
-    def __hash__(self):
-        return 0
-
-    def __bool__(self):
-        return False
-
-
-MISSING = _Missing()
+__all__ = (
+    "DictSerializerMixin",
+    "ClientSerializerMixin",
+)
 
 
 @attrs.define(eq=False, init=False, on_setattr=attrs.setters.NO_OP)
